Make it possible to not stop execution after a given amount of steps
If I am not mistaken, there is currently no way to disable the "stop
after N steps" feature. This commit allows disabling this feature
by passing a zero value to the `-s` option. This is best combined
with the `--halt-pc` option. I also think that `-s 0` would be a better
default than `-s 10000` but refrained from changing the default since
this would be a backwards-incompatible change.